refactor(client): fold the /simplify findings on the #1714 diff
The promoted-import dot specs were appended to `PeekStrip.svelte.test.ts`,
which is the outline-TICK suite; `peek-strip.test.ts` is the annotation-DOT
suite one filename-case apart, and it already owns the fixtures they were
re-building by hand. Two suites for two halves of one component is the
existing split — the new specs were simply on the wrong side of it.

`MarginColumn` resolved the display author twice per annotation per render;
the leader stroke and the two `data-tandem-author` attributes now read one
`{@const}`, which also makes it structurally impossible for them to disagree.

The remaining author-keyed sites are left on the raw field on purpose, and
now say why: `promotedAnnotation` retypes `note -> comment` in the same write
that rewrites the author, so every one of them already excludes a promoted
record through its `type` half. That invariant lived only inside
`promotedAnnotation`, which is what made the exemptions read as oversights.
