+++ tests/client/peek-strip.test.ts
@@ -66,3 +66,50 @@ describe("PeekStrip per-agent dot color (#1123 M4)", () => {
     expect(dotStyle(container, "a1")).toBeNull();
   });
 });
+
+describe("PeekStrip annotation dots — a promoted import keeps the import dot (#1714)", () => {
+  // The sliver previews the rail it is hiding. `promotedAnnotation` rewrites
+  // `author: "import" -> "user"`, so before this fix the preview dot said "the
+  // user wrote this" while the card it previews said "From: <reviewer>" — the
+  // same contradiction as the card, one surface over and far easier to miss.
+  const promoted = () =>
+    makeAnnotation({
+      id: "ann-promoted",
+      author: "user",
+      importSource: { author: "Dana Reviewer", file: "/draft.docx" },
+    });
+
+  function dotClasses(container: HTMLElement, id: string): string[] {
+    const dot = container.querySelector<HTMLElement>(`[data-testid='peek-dot-${id}']`);
+    expect(dot, `expected a dot for ${id}`).not.toBeNull();
+    return (dot?.className ?? "")
+      .split(/\s+/)
+      .filter((c) => c !== "peek-dot" && !c.startsWith("svelte-"));
+  }
+
+  it("paints the import dot for a promoted import", () => {
+    const { container } = renderStrip([promoted()]);
+    expect(dotClasses(container, "ann-promoted")).toEqual(["import"]);
+  });
+
+  it("still paints the user dot for an ordinary user comment", () => {
+    // The negative control: without it, "always return import" passes above.
+    const { container } = renderStrip([makeAnnotation({ id: "ann-plain", author: "user" })]);
+    expect(dotClasses(container, "ann-plain")).toEqual(["user"]);
+  });
+
+  it("keeps type outranking author — a promoted import carrying a suggestion reads 'suggest'", () => {
+    // `dotClass` checks highlight and suggestion BEFORE author, and threading
+    // the display author through must not disturb that order: a suggestion is
+    // what the reader acts on, and "import" would bury it.
+    const { container } = renderStrip([
+      makeAnnotation({
+        id: "ann-suggest",
+        author: "user",
+        suggestedText: "replacement",
+        importSource: { author: "Dana Reviewer", file: "/draft.docx" },
+      }),
+    ]);
+    expect(dotClasses(container, "ann-suggest")).toEqual(["suggest"]);
+  });
+});
